fix author url bug
Fixes a bug introduced in commit dae866b.

In the AUTHOR section,
e-mails automatically get converted to mailto: links.
Also the "author name <author@address>" line
is sometimes followed by an "(https://author.tld)" line
which is our magic format to turn the previous line into a labeled link.

There was an exception in the output postprocessing fn
to not do that when the previous line ends with a closing angle bracket (>)
because it indicates an e-mail link
but since commit dae866b we replace closing angle brackts with the "&gt;" entity.

This commit extends the exception so it will recognize "&gt;" as well.